# fix(p2p): Increase the size of the kad mem store