This allows adding large files without getting an error. It is a stopgap until we have a persistent kad store.

Before:
```
❯ time cargo run --release -p iroh-ctl -- add --no-wrap test1g  
    Finished release [optimized] target(s) in 0.17s
     Running `target/release/iroh add --no-wrap test1g`
Error: status: Internal, message: "the store cannot contain any more provider records", details: [], metadata: MetadataMap { headers: {"content-type": "application/grpc", "date": "Tue, 11 Oct 2022 10:39:21 GMT", "content-length": "0"} }
cargo run --release -p iroh-ctl -- add --no-wrap test1g  0.71s user 0.33s system 7% cpu 13.072 total
```

After:
```
❯ time cargo run --release -p iroh-ctl -- add --no-wrap test1g
    Finished release [optimized] target(s) in 0.18s
     Running `target/release/iroh add --no-wrap test1g`
/ipfs/bafybeiepocidrgivy3uqfh54klmtmfjolx7mceouvt2cggrpkghwsdxzni
cargo run --release -p iroh-ctl -- add --no-wrap test1g  4.75s user 2.29s system 4% cpu 2:53.84 total
```