@dignifiedquire 2022-10-11
this will still fail if you do it many times asfaik. so this is fine to improve things for the moment but we very much need to store provider records on disk to avoid this limitation entirely