@rklaehn 2022-10-11
> this will still fail if you do it many times asfaik. so this is fine to improve things for the moment but we very much need to store provider records on disk to avoid this limitation entirely

Yes, that would be the very next thing I would do. Just wanted to do something quickly so you can at least add a 1gb file without error.